src/modules/video_construction/requirements.txt
ffmpeg-python>=0.2.0
-e ../../utilities
